Refactor components and map meetup array
